// frontend/src/app/models/enum/german-state.enum.ts
export enum GermanStateEnum {
  BadenWuerttemberg = "BADEN_WUERTTEMBERG",
  Bayern = "BAYERN",
  Berlin = "BERLIN",
  Brandenburg = "BRANDENBURG",
  Bremen = "BREMEN",
  Hamburg = "HAMBURG",
  Hessen = "HESSEN",
  MecklenburgVorpommern = "MECKLENBURG_VORPOMMERN",
  Niedersachsen = "NIEDERSACHSEN",
  NordrheinWestfalen = "NORDRHEIN_WESTFALEN",
  RheinlandPfalz = "RHEINLAND_PFALZ",
  Saarland = "SAARLAND",
  Sachsen = "SACHSEN",
  SachsenAnhalt = "SACHSEN_ANHALT",
  SchleswigHolstein = "SCHLESWIG_HOLSTEIN",
  Thueringen = "THUERINGEN"
}